refactor(volList): вынес работу с фильтрами в отдельный хук
