+++ packages/admin/src/components/entities/vols/index.ts
@@ -2,3 +2,4 @@ export * from './list';
 export * from './create';
 export * from './edit';
 export * from './show';
+export { useFilters } from '~/components/entities/vols/vol-list/use-filters';
